- add htons() and ntohs() for seqno, we want endianess compatibility!! - change originator_interval to unsigned long which makes no difference on 32bit, but   removes nasty compiler warnings when checking for LONG_MIN and LONG_MAX. (long is 64bit on amd64)
git-svn-id: http://downloads.open-mesh.org/svn/batman/trunk/batman@216 45894c77-fb22-0410-b583-ff6e7d5dbf6c

+++ batman.c
@@ -82,7 +82,7 @@ short gateway_class = 0;
 short routing_class = 0;
 
 
-unsigned int orginator_interval = 1000;   /* orginator message interval in miliseconds */
+unsigned long orginator_interval = 1000;   /* orginator message interval in miliseconds */
 
 unsigned int bidirectional_timeout = 0;   /* bidirectional neighbour reply timeout in ms */
